Fix older/newer post buttons not disabling

+++ src/Actions/PostListAction.php
@@ -64,7 +64,7 @@ final class PostListAction implements RequestHandlerInterface
             [
                 'posts' => $posts,
                 'current_page' => $page,
-                'total_pages' => max(ceil($total / $limit), 1),
+                'total_pages' => (int) max(ceil($total / $limit), 1),
             ]
         );
     }
